created abstract class 'ComponentCard', added inheritance from it for other classes
