+++ src/card/card.js
@@ -1,26 +1,19 @@
-import {
-  createElement
-} from '../utils.js';
+import ComponentCard from './component-card.js';
 import {
   getTemplate
 } from './createCardTemplate.js';
 
 
-export default class Card {
+export default class Card extends ComponentCard {
   constructor(data) {
-    this._data = data;
+    super(data);
 
-    this._element = null;
     this._onClickInjected = null;
     this._onClick = this._onClick.bind(this);
   }
 
   get template() {
     return getTemplate(this._data);
-  }
-
-  get element() {
-    return this._element;
   }
 
   get id() {
@@ -35,19 +28,8 @@ export default class Card {
     this._element.addEventListener(`click`, this._onClick);
   }
 
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
   unbind() {
     this._element.removeEventListener(`click`, this._onClick);
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
   }
 
   _onClick(event) {
